e2e/ai-tests/requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
backoff==2.2.1
beautifulsoup4==4.13.3
browser-use==0.1.40
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
defusedxml==0.7.1
distro==1.9.0
eval_type_backport==0.2.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-anthropic==0.3.3
langchain-core==0.3.49
langchain-ollama==0.2.2
langchain-openai==0.3.1
langsmith==0.3.19
markdownify==0.14.1
mistralai==1.6.0
monotonic==1.6
ollama==0.4.7
openai==1.68.2
orjson==3.10.16
packaging==24.2
playwright==1.51.0
posthog==3.23.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
zstandard==0.23.0